+++ AsBuiltReport.Veeam.VBR/Src/Private/Diagram/Get-AbrTapeVaultInfo.ps1
@@ -21,7 +21,7 @@ function Get-AbrTapeVaultInfo {
         Retrieves and displays information about all Tape Vaults from the Veeam Backup & Replication server.
 
     .NOTES
-        Author: Jonathan Colon
+        Author: AsBuiltReport
         Date: 2024-12-31
         Version: 1.0
     #>