+++ AsBuiltReport.Veeam.VBR/Src/Private/Diagram/Get-AbrVbrServerConnection.ps1
@@ -13,14 +13,14 @@ function Get-AbrServerConnection {
         The TCP Port of the target Veeam Backup Server.
 
     .NOTES
-        Version:        1.0.0
-        Author:         Jonathan Colon
-        Twitter:        @jcolonfzenpr
-        GitHub:         rebelinux
+        Version:        1.0.3
+        Author:         AsBuiltReport Organization
+        Twitter:        @asbuiltreport
+        Github:         asbuiltreport
         Credits:        Iain Brighton (@iainbrighton) - PScribo module
 
     .LINK
-        https://github.com/rebelinux/Veeam.Diagrammer
+        https://github.com/AsBuiltReport/AsBuiltReport.Veeam.VBR
 
     .EXAMPLE
         PS> Get-AbrServerConnection -Port 9392