+++ AsBuiltReport.Veeam.VBR/Src/Private/Report/Get-AbrVbrBackupServerInfo.ps1
@@ -6,10 +6,10 @@ function Get-AbrVbrBackupServerInfo {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        1.0.1
-        Author:         Jonathan Colon
-        Twitter:        @jcolonfzenpr
-        Github:         rebelinux
+        Version:        1.0.3
+        Author:         AsBuiltReport Organization
+        Twitter:        @asbuiltreport
+        Github:         asbuiltreport
         Credits:        Iain Brighton (@iainbrighton) - PScribo module
 
     .LINK
@@ -32,7 +32,7 @@ function Get-AbrVbrBackupServerInfo {
                 { $_.Major -lt 13 } { Get-VBRServer -Type Local }
                 default { Get-VBRServer | Where-Object { $_.Description -eq 'Backup server' } }
             }
-            if (($VbrVersion.Major -gt 13) -and (Get-VBRServer | Where-Object { $_.Description -eq 'Backup server' -and $_.Type -eq 'Linux' } )) {
+            if (($VbrVersion.Major -ge 13) -and (Get-VBRServer | Where-Object { $_.Description -eq 'Backup server' -and $_.Type -eq 'Linux' } )) {
                 $VeeamVersion = @{
                     DisplayVersion = $VbrVersion
                 }
@@ -578,7 +578,9 @@ function Get-AbrVbrBackupServerInfo {
                     #---------------------------------------------------------------------------------------------#
                     try {
                         Write-PScriboMessage $LocalizedData.CollectingHA
-                        $HACluster = try { Get-VBRHighAvailabilityCluster -WarningAction SilentlyContinue } catch { Out-Null }
+                        if ($VbrVersion.Major -ge 13) {
+                            $HACluster = try { Get-VBRHighAvailabilityCluster -WarningAction SilentlyContinue } catch { Out-Null }
+                        }
                         if ($HACluster) {
                             Section -Style Heading4 $LocalizedData.HAHeading {
                                 $OutObj = @()
@@ -648,7 +650,9 @@ function Get-AbrVbrBackupServerInfo {
                                 }
                                 if ($Options.EnableDiagrams -and ($VbrVersion -ge [version]'12.1')) {
                                     try {
-                                        $HAClusterCheck = try { Get-VBRHighAvailabilityCluster -WarningAction SilentlyContinue } catch { Out-Null }
+                                        if ($VbrVersion.Major -ge 13) {
+                                            $HACluster = try { Get-VBRHighAvailabilityCluster -WarningAction SilentlyContinue } catch { Out-Null }
+                                        }
                                         if ($HAClusterCheck) {
                                             try {
                                                 $Graph = Get-AbrVbrDiagrammer -DiagramType 'Backup-to-HACluster' -DiagramOutput base64