+++ AsBuiltReport.Veeam.VBR/Src/Private/Report/Get-StrdDevDuration.ps1
@@ -5,7 +5,7 @@ function Get-StrdDevDuration {
     .DESCRIPTION
     .NOTES
         Version:        0.1.0
-        Author:         Jonathan Colon
+        Author:         AsBuiltReport Organization
     .EXAMPLE
         Get-StrdDevDuration -$JobTimeSpan
     .LINK